Use Git to fetch upstream uncrustify source code
This is somewhat of a lateral move in general, but the patching solution
implemented in ec709bb8838f193dc1c90ccf95ecc1e28ba6bd08 can fail under
certain circumstances if the target directory is not a git repository.

Since the patching solution requires git anyway, and we're downloading
what should be the exact same files, there should be no reason that this
wouldn't work in largely the same way the current tarball-based solution
does.

Signed-off-by: Scott K Logan <logans@cottsay.net>