Use explicit INSTALL_COMMAND to suppress DESTDIR (#21)
The DESTDIR environment variable is used on UNIX systems during an
invocation of 'make install' to specify an alternate directory prefix
for installation paths. Since this isn't a cross-platform mechanism,
the CMAKE_INSTALL_PREFIX CMake variable can also be used.

Unfortunately, if both of these are specified, they will both be
applied. This isn't typically a problem, but can cause problems for the
installation phase of ExternalProject_Add if the external project is
built during an invocation of 'make install' with DESTDIR set.

Since the CMAKE_INSTALL_PREFIX method is employed by the call to
ExternalProject_Add instead of DESTDIR, we should specifically suppress
DESTDIR if it is set so that it doesn't interfere with the installation
of the external project to the staging directory.

Signed-off-by: Scott K Logan <logans@cottsay.net>