Make git upstream act more like tar (#22)
Some of our problems with using GIT_REPOSITORY with a GIT_TAG stem from
CMake trying to "update" the repository each time the target is
evaluated. This would be important if not for the fact that we're
targeting a specific REF, so the repository will never be out-of-date
unless we change the ref, which would invalidate the target anyway.

We can safely use UPDATE_DISCONNECTED to block the update step from
invalidating the source target, which should prevent downstream targets
like patch, build, and install from being re-triggered unnecessarily.

The one caveat to this is that a change to GIT_TAG doesn't invalidate
the target like a change to GIT_REPOSITORY or URL would. To work around
this, I updated the target name to contain the ref instead of the
supposed version number, which will then force a new build if we change
the ref in the future. This is also more correct, since it would be easy
for the ref and target name to get out of sync.

Signed-off-by: Scott K Logan <logans@cottsay.net>